+++ projects/js-packages/connection/components/connect-button/index.jsx
@@ -6,6 +6,7 @@ import { __ } from '@wordpress/i18n';
 import { Button } from '@wordpress/components';
 import PropTypes from 'prop-types';
 import restApi from '@automattic/jetpack-api';
+import { Spinner } from '@automattic/jetpack-components';
 
 /**
  * Internal dependencies
@@ -26,11 +27,13 @@ import './style.scss';
  * @param {string} props.from -- Where the connection request is coming from.
  * @param {object} props.connectionStatus -- The connection status object.
  * @param {boolean} props.connectionStatusIsFetching -- The flag indicating that connection status is being fetched.
+ * @param {boolean} props.autoTrigger -- Whether to initiate the connection process automatically upon rendering the component.
  * @returns {React.Component} The RNA connection component.
  */
 const ConnectButton = props => {
 	const [ isRegistering, setIsRegistering ] = useState( false );
 	const [ isUserConnecting, setIsUserConnecting ] = useState( false );
+	const [ registationError, setRegistrationError ] = useState( false );
 
 	const [ authorizationUrl, setAuthorizationUrl ] = useState( null );
 
@@ -44,6 +47,7 @@ const ConnectButton = props => {
 		from,
 		connectionStatus,
 		connectionStatusIsFetching,
+		autoTrigger,
 	} = props;
 
 	/**
@@ -60,6 +64,8 @@ const ConnectButton = props => {
 	const registerSite = useCallback(
 		e => {
 			e && e.preventDefault();
+
+			setRegistrationError( false );
 
 			if ( connectionStatus.isRegistered ) {
 				setIsUserConnecting( true );
@@ -82,6 +88,7 @@ const ConnectButton = props => {
 				} )
 				.catch( error => {
 					setIsRegistering( false );
+					setRegistrationError( error );
 					throw error;
 				} );
 		},
@@ -94,6 +101,15 @@ const ConnectButton = props => {
 			redirectUri,
 		]
 	);
+
+	/**
+	 * Auto-trigger the flow, only do it once.
+	 */
+	useEffect( () => {
+		if ( autoTrigger && ! isRegistering && ! isUserConnecting ) {
+			registerSite();
+		}
+	}, [] ); // eslint-disable-line react-hooks/exhaustive-deps
 
 	return (
 		<div className="jp-connect-button">
@@ -108,9 +124,15 @@ const ConnectButton = props => {
 						isPrimary
 						disabled={ isRegistering || isUserConnecting }
 					>
-						{ connectLabel }
+						{ isRegistering || isUserConnecting ? <Spinner /> : connectLabel }
 					</Button>
 				) }
+
+			{ registationError && (
+				<p className="jp-connect-button__error">
+					{ __( 'An error occurred. Please try again.', 'jetpack' ) }
+				</p>
+			) }
 
 			{ isUserConnecting && (
 				<ConnectUser connectUrl={ authorizationUrl } redirectUri={ redirectUri } from={ from } />
@@ -127,11 +149,13 @@ ConnectButton.propTypes = {
 	from: PropTypes.string,
 	redirectUri: PropTypes.string.isRequired,
 	registrationNonce: PropTypes.string.isRequired,
+	autoTrigger: PropTypes.bool,
 };
 
 ConnectButton.defaultProps = {
 	connectLabel: __( 'Connect', 'jetpack' ),
 	redirectUri: null,
+	autoTrigger: false,
 };
 
 export default ConnectButton;