+++ projects/plugins/jetpack/modules/masterbar/admin-menu/class-base-admin-menu.php
@@ -78,7 +78,7 @@ abstract class Base_Admin_Menu {
 			add_action( 'admin_head', array( $this, 'set_site_icon_inline_styles' ) );
 			add_filter( 'screen_settings', array( $this, 'register_dashboard_switcher' ), 99999 );
 			add_action( 'admin_menu', array( $this, 'handle_preferred_view' ), 99997 );
-			add_action( 'wp_ajax_set_preferred_view', array( $this, 'handle_preferred_view' ) );
+			add_action( 'wp_ajax_set_preferred_view', array( $this, 'handle_preferred_view_ajax' ) );
 		}
 	}
 
@@ -88,7 +88,7 @@ abstract class Base_Admin_Menu {
 	 * @return Admin_Menu
 	 */
 	public static function get_instance() {
-		$class = get_called_class();
+		$class = static::class;
 
 		if ( empty( static::$instances[ $class ] ) ) {
 			static::$instances[ $class ] = new $class();
@@ -612,7 +612,7 @@ abstract class Base_Admin_Menu {
 		if ( isset( $_GET['page'] ) ) {
 			$screen = add_query_arg( 'page', $_GET['page'], $screen );
 		}
-		return $screen;
+		return sanitize_text_field( wp_unslash( $screen ) );
 		// phpcs:enable WordPress.Security.NonceVerification
 	}
 
@@ -647,11 +647,15 @@ abstract class Base_Admin_Menu {
 		 * @param string The preferred view the user selected.
 		 */
 		\do_action( 'jetpack_dashboard_switcher_changed_view', $current_screen, $preferred_view );
-
-		if ( wp_doing_ajax() ) {
-			wp_die();
-		}
 		// phpcs:enable WordPress.Security.NonceVerification
+	}
+
+	/**
+	 * Handles AJAX requests setting a preferred view for a given screen.
+	 */
+	public function handle_preferred_view_ajax() {
+		$this->handle_preferred_view();
+		wp_die();
 	}
 
 	/**