+++ projects/plugins/jetpack/modules/search/customberg/components/header/index.jsx
@@ -1,42 +1,42 @@
 /**
  * WordPress dependencies
  */
-import { VisuallyHidden } from '@wordpress/components';
-import { PinnedItems } from '@wordpress/interface';
+import { Button } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
-import { useViewportMatch } from '@wordpress/compose';
+import { Icon, cog } from '@wordpress/icons';
 
 /**
  * Internal dependencies
  */
-import { COMPLEMENTARY_AREA_SCOPE } from '../../lib/constants';
 import SaveButton from './save-button';
 import './styles.scss';
 
 /**
  * Component for showing the Gutenberg-style header.
  *
+ * @param {object} props - component properties.
+ * @param {Function} props.enableSidebar - Enables the sidebar upon invocation.
  * @returns {Element} component instance
  */
-function Header() {
-	const isMediumViewport = useViewportMatch( 'medium' );
+function Header( { enableSidebar } ) {
+	// translators: Product name 'Jetpack Search' should not be translated
+	const title = __( 'Customize Jetpack Search', 'jetpack' );
 
 	return (
 		<div className="jp-search-configure-header">
 			<div className="jp-search-configure-header__navigable-toolbar-wrapper">
-				{ isMediumViewport ? (
-					<h1 className="jp-search-configure-header__title">
-						{ __( 'Jetpack Search', 'jetpack' ) }
-					</h1>
-				) : (
-					<VisuallyHidden as="h1" className="jp-search-configure-header__title">
-						{ __( 'Jetpack Search', 'jetpack' ) }
-					</VisuallyHidden>
-				) }
+				<h1 className="jp-search-configure-header__title">{ title }</h1>
 			</div>
 			<div className="jp-search-configure-header__actions">
 				<SaveButton />
-				<PinnedItems.Slot scope={ COMPLEMENTARY_AREA_SCOPE } />
+				<Button
+					aria-label={ __( 'Show settings', 'jetpack' ) }
+					className="jp-search-configure-header__show-settings-button"
+					isSecondary
+					onClick={ () => enableSidebar() }
+				>
+					<Icon icon={ cog } />
+				</Button>
 			</div>
 		</div>
 	);