+++ projects/plugins/jetpack/tests/e2e/lib/pages/wpcom/connections.js
@@ -19,19 +19,16 @@ export default class ConnectionsPage extends WpPage {
 	async selectMailchimpList( mailchimpList = 'e2etesting' ) {
 		const loadingIndicatorSelector = '.foldable-card__summary button:not([disabled])';
 		const mailchimpExpandSelector = '.mailchimp .foldable-card__expand';
-		const marketingSelectSelector = '.mailchimp select';
+		const marketingSelectSelector = '.mailchimp.is-expanded select ';
 		const mcOptionXpathSelector = `//option[contains(text(), '${ mailchimpList }')]`;
 		const successNoticeSelector = `//span[contains(text(), '${ mailchimpList }')]`;
 
 		await this.waitForElementToBeVisible( loadingIndicatorSelector );
-
 		await this.click( mailchimpExpandSelector );
 
-		// WPCOM Connections page
 		await this.waitForElementToBeAttached( mcOptionXpathSelector );
 		await this.selectOption( marketingSelectSelector, { label: mailchimpList } );
-		await this.waitForElementToBeVisible( successNoticeSelector );
-		await this.page.close();
+		await this.waitForElementToBeVisible( successNoticeSelector, 10000 );
 	}
 
 	async connectMailchimp() {