+++ projects/packages/changelogger/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2021-08-30
+### Changed
+- Run composer update on test-php command instead of phpunit
+- Tests: update PHPUnit polyfills dependency (yoast/phpunit-polyfills).
+
 ## [1.2.0] - 2021-05-12
 ### Added
 - New option, `--filename-auto-suffix`, to ensure that a reused branch won't prevent entry creation in non-interactive mode.
@@ -36,6 +41,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 - Initial version.
 
+[1.2.1]: https://github.com/Automattic/jetpack-changelogger/compare/1.2.0...1.2.1
 [1.2.0]: https://github.com/Automattic/jetpack-changelogger/compare/1.1.2...1.2.0
 [1.1.2]: https://github.com/Automattic/jetpack-changelogger/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/Automattic/jetpack-changelogger/compare/1.1.0...1.1.1